release(agent): wire vp9-444 into Linux/macOS/Windows MSI release pipeline
Splits env-libvpx-sys's `generate` (bindgen at build time) into a
separate `vp9-444-bindgen` cargo feature so the Windows release path
can opt out of bindgen entirely.

Why: bindgen 0.68.1 + libclang on Windows has a regression that
opaques `vpx_codec_enc_cfg` (forward-declared via pointer-to-incomplete
in vpx_codec.h's union before its full def in vpx_encoder.h) — every
field reference becomes `_address: u8` and the build collapses. Local
repro on libvpx 1.12.0 + 1.13.1 + 1.16.0 + LLVM 22.1.4. Linux + macOS
bindgen handles it correctly.

Workaround: skip bindgen on Windows. env-libvpx-sys ships pre-generated
FFI for libvpx 1.3.0–1.13.0; the Windows release pins libvpx to 1.12.0
via vcpkg so the pre-generated bindings match the linked lib. Linux +
macOS keep `vp9-444-bindgen` (apt's libvpx 1.14, brew's libvpx 1.15
both work fine through bindgen).

Also fix `frame_pkt.sz as usize` cast — the field is `u32` in the
1.12.0 pre-generated binding (Windows path) and `usize` in 1.13.x +
bindgen-generated bindings (Linux/macOS path).

.deb runtime depend pruned to `libvpx9 | libvpx7` (apt provides the
shared lib; libvpx-dev is the devel headers package and not needed at
runtime).

+++ agents/roomler-agent/src/encode/libvpx.rs
@@ -409,7 +409,14 @@ impl VideoEncoder for Vp9Encoder {
             // SAFETY: the union variant is `frame` for FRAME_PKT.
             let frame_pkt = unsafe { (*pkt).data.frame };
             let buf = frame_pkt.buf as *const u8;
-            let sz = frame_pkt.sz;
+            // `sz` field width varies across env-libvpx-sys's
+            // pre-generated bindings (u32 on the 1.12 binding used by
+            // the Windows release path; usize on 1.13 + the
+            // bindgen-against-system-headers path used on Linux/macOS).
+            // Cast through usize unconditionally — the value is a byte
+            // count of an encoder packet, no realistic risk of overflow
+            // on 64-bit hosts.
+            let sz = frame_pkt.sz as usize;
             // SAFETY: libvpx guarantees the buffer is at least `sz`
             // bytes long and stays valid until the next encode() call.
             let slice = unsafe { std::slice::from_raw_parts(buf, sz) };
